Use CMake to build static libraries for Android
Currently, there are a couple of issues in building the Android app:

1. Toolchain too old: This includes both NDK, Android SDK, Android Build Tools, Android Gradle Plugin, and Kotlin

2. Libraries are totally dynamically linked. This creates issues with libzstd because it hard-codes SONAME [1], while the Android NDK linker ignores SONAME symlinks (libzstd.so -> libzstd.so.1, SONAME = libzstd.so.1 WILL result in libzstd.so (regular file), SONAME = libzstd.so.1), and ld.so will not able to find that library [2]. The previous approach was to use patchelf(1) to change the SONAME[3], but the so file will be striped using llvm-strip(1) as part of the Android NDK build system, and the resulting ELF is broken [4].

Thus, the best solution is to statically link everything. Even better, use CMake to find the libraries and headers instead of hard-coding the /sdr-kit/ in CMakeLists.txt, since most libraries are providing CMake xxxconfig.cmake files, and we can easily provide our own for the absent ones.

This patchset was tested using the mentioned Android SDK, NDK, and CMake versions as shown in build.gradle. I did not yet test them in Docker because I do not use Docker personally.

This patchset was not tested against normal desktop builds. I guess it won't cause any harm because I didn't touch parts outside of ifdef ANDROID.

A subsequent patch will be sent to the android-sdr-kit.git to reflect the necessary changes

With this patch, dynamic libraries had been cut down to only a few, and there was no libzstd.so issues anymore.

[1]: https://github.com/facebook/zstd/blob/82d470564d01c42ee62d651e644c5ea6953561e2/lib/Makefile#L76-L90

[2]: https://groups.google.com/g/android-ndk/c/_UhNpRJlA1k/m/hbryqzEgN94J

[3]: https://github.com/AlexandreRouma/android-sdr-kit/blob/6589435134e446dfa6c007cf79e8557d117f3702/build.sh#L90

[4]: https://github.com/NixOS/patchelf/issues/507

Signed-off-by: Yuuta Liang <yuuta@yuuta.moe>